# PERF: don't build coincident lookup if not needed