--- a/libpysal/graph/_triangulation.py
+++ b/libpysal/graph/_triangulation.py
@@ -58,10 +58,11 @@ def _validate_coincident(triangulator):
         )
 
         # check for coincident points
-        n_coincident, coincident_lut = _build_coincidence_lookup(geoms)
+        n_coincident = geoms.geometry.duplicated().sum()
 
         # resolve coincident points prior triangulation
         if n_coincident > 0:
+            coincident_lut = _build_coincidence_lookup(geoms)
             if coincident == "raise":
                 raise ValueError(
                     f"There are {len(coincident_lut)} unique locations in "
@@ -129,7 +130,7 @@ def _validate_coincident(triangulator):
             coordinates, ids, geoms = input_coordinates, input_ids, input_geoms
             heads, tails, weights = adjtable.values.T
 
-        adjtable = _reorder_adjtable_by_ids(adjtable, ids)
+            adjtable = _reorder_adjtable_by_ids(adjtable, ids)
 
         # return data for Graph.from_arrays
         return adjtable.focal.values, adjtable.neighbor.values, adjtable.weight.values